TileTest Update
The quest for ultimate Tiled Integration continues. In this exciting
episode, the implementation of Properties was used to make it so that
you can't walk across walls and fire will oxymoronically turn you into a
bomb. As always, simply modify the AndroidLauncher file to run TileTest
instead of the actual game.